Renamed: 'ViewModInfo' to 'ModInfo'
