+++ src/NexusMods.App.UI/Pages/ModInfo/IModInfoViewModel.cs
@@ -1,15 +1,15 @@
 using NexusMods.Abstractions.Loadouts;
 using NexusMods.Abstractions.Loadouts.Mods;
-using NexusMods.App.UI.Pages.ViewModInfo.Types;
+using NexusMods.App.UI.Pages.ModInfo.Types;
 using NexusMods.App.UI.WorkspaceSystem;
 
-namespace NexusMods.App.UI.Pages.ViewModInfo;
+namespace NexusMods.App.UI.Pages.ModInfo;
 
-public interface IViewModInfoViewModel : IPageViewModelInterface
+public interface IModInfoViewModel : IPageViewModelInterface
 { 
     LoadoutId LoadoutId { get; set; }
     ModId ModId { get; set; }
-    CurrentViewModInfoPage Page { get; set; }
+    CurrentModInfoPage Page { get; set; }
     IViewModelInterface PageViewModel { get; set; }
 
     /// <summary>
