+++ src/NexusMods.App.UI/Pages/ModInfo/ModInfoView.axaml.cs
@@ -1,0 +1,19 @@
+﻿using System.Reactive.Disposables;
+using Avalonia.ReactiveUI;
+using ReactiveUI;
+
+namespace NexusMods.App.UI.Pages.ModInfo;
+
+public partial class ModInfoView : ReactiveUserControl<IModInfoViewModel>
+{
+    public ModInfoView()
+    {
+        InitializeComponent();
+        this.WhenActivated(d =>
+        {
+            this.OneWayBind(ViewModel, vm => vm.PageViewModel, view => view.ModInfoCategoryViewHost.ViewModel)
+                .DisposeWith(d);
+        });
+    }
+}
+
