+++ src/NexusMods.App.UI/Pages/ModInfo/Types/CurrentModInfoPage.cs
@@ -1,0 +1,12 @@
+namespace NexusMods.App.UI.Pages.ModInfo.Types;
+
+/// <summary>
+///     Currently shown page in <see cref="ModInfoView"/>.
+/// </summary>
+public enum CurrentModInfoPage
+{
+    /// <summary>
+    ///     For the file tree.
+    /// </summary>
+    Files,
+}
