--- src/NexusMods.App.UI/Pages/ViewModInfo/Types/CurrentViewModInfoPage.cs
@@ -1,12 +1,0 @@
-namespace NexusMods.App.UI.Pages.ViewModInfo.Types;
-
-/// <summary>
-///     Currently shown page in <see cref="ViewModInfoView"/>.
-/// </summary>
-public enum CurrentViewModInfoPage
-{
-    /// <summary>
-    ///     For the file tree.
-    /// </summary>
-    Files,
-}
