--- src/NexusMods.App.UI/Pages/ViewModInfo/ViewModInfoView.axaml.cs
@@ -1,19 +1,0 @@
-﻿using System.Reactive.Disposables;
-using Avalonia.ReactiveUI;
-using ReactiveUI;
-
-namespace NexusMods.App.UI.Pages.ViewModInfo;
-
-public partial class ViewModInfoView : ReactiveUserControl<IViewModInfoViewModel>
-{
-    public ViewModInfoView()
-    {
-        InitializeComponent();
-        this.WhenActivated(d =>
-        {
-            this.OneWayBind(ViewModel, vm => vm.PageViewModel, view => view.PageViewHost.ViewModel)
-                .DisposeWith(d);
-        });
-    }
-}
-
