+++ src/NexusMods.App.UI/TypeFinder.cs
@@ -2,8 +2,8 @@ using NexusMods.Abstractions.Serialization.ExpressionGenerator;
 using NexusMods.App.UI.Controls;
 using NexusMods.App.UI.Pages.Downloads;
 using NexusMods.App.UI.Pages.LoadoutGrid;
+using NexusMods.App.UI.Pages.ModInfo;
 using NexusMods.App.UI.Pages.MyGames;
-using NexusMods.App.UI.Pages.ViewModInfo;
 using NexusMods.App.UI.WorkspaceSystem;
 
 namespace NexusMods.App.UI;
